⚙️ feat: 하단 navigating tab 수정 및 온보딩 스크린 로직 추가
✅ Closes: #42
