+++ screens/BottomTabs/index.js
@@ -1,0 +1,3 @@
+import BottomTabs from './BottomTabs';
+
+export default BottomTabs;
