+++ screens/Home/Home.jsx
@@ -1,7 +1,5 @@
-import { useNavigation } from '@react-navigation/native';
 import React from 'react';
-import { FlatList, ScrollView, Text, TouchableOpacity } from 'react-native';
-import Button from '@components/Button';
+import { ScrollView, Text } from 'react-native';
 import TrendingTags from '@containers/home/TrendingTags';
 import RecommendedRoutes from '@containers/home/RecommendedRoutes';
 import RecommendedRoutesNearBy from '@containers/home/RecommendedRoutesNearBy';
@@ -16,12 +14,6 @@ const Home = ({ navigation }) => {
         <TrendingTags />
         <RecommendedRoutesNearBy />
       </ScrollView>
-      <Button
-        title={'경로 기록하기'}
-        onPress={() => {
-          navigation.navigate('Recording');
-        }}
-      />
     </S.Wrapper>
   );
 };
