+++ screens/Onboarding/Onboarding.jsx
@@ -1,12 +1,14 @@
 import React from 'react';
-import * as S from './Onboarding.style.js';
+import Button from '@components/Button';
+import * as S from './OnBoarding.style.js';
 
-const Onboarding = () => {
+const OnBoarding = ({ navigation }) => {
   return (
     <S.Wrapper>
       <S.Text>Onboarding Screen</S.Text>
+      <Button title={'Home으로 이동'} onPress={() => navigation.navigate('afterOnboarding')} />
     </S.Wrapper>
   );
 };
 
-export default Onboarding;
+export default OnBoarding;
